[refactor]: Refactoring and removed the duplicate code
The project is reviewed as an objective for removing the duplicate code,
so as to optimize the maintenance  as initially the database, messageBus
to connect to the queue, spider and more was same duplicate files,
which is currently changed to packages.

Moved the main applications go_crawler and go_supervisor to the
/cmd folder. It is common to have a small main function that imports
and invokes the code from the /internal and other package directories as
dependencies.

The make file for provision is also refactored as the docker files are
all moved to the docker directory.

The file naming conventions is changed to common practice as seen in
most of the repositories (messageBus.go to message_bus.go).

🍻